Implemented routing for homepage and swipe page, renamed UserList(container) to DineOut(container)
