+++ src/components/DineOutList.js
@@ -1,0 +1,11 @@
+//components/DineOutList.js
+import * as React from 'react'
+
+export default function DineOutList(props) {
+    return (
+        <div>
+            You're now in component: DineOutList
+            This is where you'll be swiping.
+        </div> 
+    )
+}
