--- src/components/UserList.js
@@ -1,22 +1,0 @@
-//components/UserList.js
-import * as React from 'react'
-
-export default function UserList(props) {
-    return (
-        <div>
-        Snoep
-
-            <div>
-            <button id="dine-out">
-            Dine out
-            </button>
-            </div>                      <br/><br/><br/><br/>
-            <div>
-            <button id="dine-in">
-            Dine in
-            </button>
-            </div>
-            
-        </div> 
-    )
-}
