+++ src/index.js
@@ -5,10 +5,14 @@ import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 import store from './store'
 import {Provider} from 'react-redux'
+import {BrowserRouter} from 'react-router-dom' 
 
 ReactDOM.render(
     <Provider store={store}>
-        <App />     
+        <BrowserRouter>
+            <App />   
+        </BrowserRouter>  
     </Provider>
+    
 , document.getElementById('root'));
 registerServiceWorker();
